requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiolimiter==1.2.1
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.60.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cryptography==46.0.3
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.1
faiss-cpu==1.7.4
fastapi==0.124.0
ffmpy==1.0.0
filelock==3.20.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
gradio==4.29.0
gradio_client==0.16.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain-core==1.1.2
langchain-text-splitters==1.0.0
langsmith==0.4.56
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
narwhals==2.13.0
networkx==3.4.2
numpy==1.26.0
orjson==3.11.5
packaging==24.2
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==10.4.0
propcache==0.4.1
protobuf==5.29.5
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pypdf==4.2.0
pypdfium2==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.30.0
ruff==0.14.8
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==5.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.50.0
streamlit==1.38.0
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.12.0
torch==2.9.1
tornado==6.5.2
tqdm==4.66.0
transformers==4.57.3
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
uuid_utils==0.12.0
uvicorn==0.38.0
voyageai==0.3.5
watchdog==4.0.2
websockets==11.0.3
yarl==1.22.0
zstandard==0.25.0
